sound150 v11.5.0: Best album cover change
