+++ sound150@claudiux/files/sound150@claudiux/metadata.json
@@ -5,7 +5,7 @@
   "max-instances": "1",
   "description": "Enhanced sound applet",
   "hide-configuration": false,
-  "version": "11.4.0",
+  "version": "11.5.0",
   "cinnamon-version": [
     "2.8",
     "3.0",
